# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/maintenance_requests:
    get:
      tags:
        - MaintenanceRequests
      summary: List MaintenanceRequests
      description: Get a paginated list of MaintenanceRequests with optional filtering and sorting
      operationId: listMaintenanceRequests
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRequestList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - MaintenanceRequests
      summary: Create MaintenanceRequest
      description: Create a new MaintenanceRequest
      operationId: createMaintenanceRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMaintenanceRequestRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    MaintenanceRequest:
      type: object
      required:
        - id
        - company_id
        - name
        - request_date
        - duration
        - stage_id
        - kanban_state
        - priority
        - maintenance_type
        - recurring
        - repeat_interval
        - repeat_unit
        - repeat_type
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          nullable: true
          type: string
        request_date:
          type: string
          format: date
        schedule_date:
          nullable: true
          type: string
          format: date-time
        schedule_end:
          nullable: true
          type: string
          format: date-time
        close_date:
          nullable: true
          type: string
          format: date
        duration:
          type: string
        owner_user_id:
          nullable: true
          type: string
          format: uuid
        user_id:
          nullable: true
          type: string
          format: uuid
        asset_id:
          nullable: true
          type: string
          format: uuid
        stage_id:
          type: string
          format: uuid
        kanban_state:
          type: string
        priority:
          type: string
        maintenance_type:
          type: string
        recurring:
          type: boolean
        repeat_interval:
          type: integer
          format: int32
        repeat_unit:
          type: string
        repeat_type:
          type: string
        repeat_until:
          nullable: true
          type: string
          format: date
        successor_request_id:
          nullable: true
          type: string
          format: uuid
        successor_of_request_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateMaintenanceRequestRequest:
      type: object
      required:
        - company_id
        - name
        - request_date
        - duration
        - stage_id
        - kanban_state
        - priority
        - maintenance_type
        - recurring
        - repeat_interval
        - repeat_unit
        - repeat_type
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          nullable: true
          type: string
        request_date:
          type: string
          format: date
        schedule_date:
          nullable: true
          type: string
          format: date-time
        schedule_end:
          nullable: true
          type: string
          format: date-time
        close_date:
          nullable: true
          type: string
          format: date
        duration:
          type: string
        owner_user_id:
          nullable: true
          type: string
          format: uuid
        user_id:
          nullable: true
          type: string
          format: uuid
        asset_id:
          nullable: true
          type: string
          format: uuid
        stage_id:
          type: string
          format: uuid
        kanban_state:
          type: string
        priority:
          type: string
        maintenance_type:
          type: string
        recurring:
          type: boolean
        repeat_interval:
          type: integer
          format: int32
        repeat_unit:
          type: string
        repeat_type:
          type: string
        repeat_until:
          nullable: true
          type: string
          format: date
        successor_request_id:
          nullable: true
          type: string
          format: uuid
        successor_of_request_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    MaintenanceRequestList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MaintenanceRequest'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []