# src/presentation/http/maintenance_request_handler.rs
//! MaintenanceRequest REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc, NaiveDate};
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{MaintenanceRequestService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateMaintenanceRequestDto, UpdateMaintenanceRequestDto, PatchMaintenanceRequestDto, MaintenanceRequestResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum MaintenanceRequestError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
    // Domain-specific errors from hook rules
    #[error("A request cannot be scheduled to finish before it starts.: {0}")]
    RequestScheduleEndBeforeStart(String),
    #[error("A recurrence step must be at least one unit.: {0}")]
    RequestRepeatIntervalBelowOne(String),
    #[error("A recurrence bounded 'until' must state its repeat_until date.: {0}")]
    RequestRepeatUntilMissing(String),
    #[error("Only preventive maintenance may recur.: {0}")]
    RequestCorrectiveCannotRecur(String),
    #[error("Stage transitions are trigger-consistent; closing a recurring preventive request requires the managed transition verb.: {0}")]
    MaintenanceRecurringCloseRequiresServiceVerb(String),
    #[error("A stage referenced by live requests cannot be deleted.: {0}")]
    MaintenanceStageInUse(String),
    #[error("The stage foreign key is real, ON DELETE RESTRICT.: {0}")]
    RequestStageFk(String),
}

impl From<ServiceError> for MaintenanceRequestError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for MaintenanceRequestError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "MAINTENANCEREQUEST_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "MAINTENANCEREQUEST_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "MAINTENANCEREQUEST_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "MAINTENANCEREQUEST_INTERNAL_ERROR"),
            Self::RequestScheduleEndBeforeStart(_) => (StatusCode::UNPROCESSABLE_ENTITY, "MAINTENANCEREQUEST_REQUEST_SCHEDULE_END_BEFORE_START"),
            Self::RequestRepeatIntervalBelowOne(_) => (StatusCode::UNPROCESSABLE_ENTITY, "MAINTENANCEREQUEST_REQUEST_REPEAT_INTERVAL_BELOW_ONE"),
            Self::RequestRepeatUntilMissing(_) => (StatusCode::UNPROCESSABLE_ENTITY, "MAINTENANCEREQUEST_REQUEST_REPEAT_UNTIL_MISSING"),
            Self::RequestCorrectiveCannotRecur(_) => (StatusCode::UNPROCESSABLE_ENTITY, "MAINTENANCEREQUEST_REQUEST_CORRECTIVE_CANNOT_RECUR"),
            Self::MaintenanceRecurringCloseRequiresServiceVerb(_) => (StatusCode::UNPROCESSABLE_ENTITY, "MAINTENANCEREQUEST_MAINTENANCE_RECURRING_CLOSE_REQUIRES_SERVICE_VERB"),
            Self::MaintenanceStageInUse(_) => (StatusCode::UNPROCESSABLE_ENTITY, "MAINTENANCEREQUEST_MAINTENANCE_STAGE_IN_USE"),
            Self::RequestStageFk(_) => (StatusCode::UNPROCESSABLE_ENTITY, "MAINTENANCEREQUEST_REQUEST_STAGE_FK"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

/// Domain-specific error codes for MaintenanceRequest
pub mod maintenance_request_errors {
    pub const REQUEST_SCHEDULE_END_BEFORE_START: &str = "MAINTENANCEREQUEST_REQUEST_SCHEDULE_END_BEFORE_START";
    pub const REQUEST_REPEAT_INTERVAL_BELOW_ONE: &str = "MAINTENANCEREQUEST_REQUEST_REPEAT_INTERVAL_BELOW_ONE";
    pub const REQUEST_REPEAT_UNTIL_MISSING: &str = "MAINTENANCEREQUEST_REQUEST_REPEAT_UNTIL_MISSING";
    pub const REQUEST_CORRECTIVE_CANNOT_RECUR: &str = "MAINTENANCEREQUEST_REQUEST_CORRECTIVE_CANNOT_RECUR";
    pub const MAINTENANCE_RECURRING_CLOSE_REQUIRES_SERVICE_VERB: &str = "MAINTENANCEREQUEST_MAINTENANCE_RECURRING_CLOSE_REQUIRES_SERVICE_VERB";
    pub const MAINTENANCE_STAGE_IN_USE: &str = "MAINTENANCEREQUEST_MAINTENANCE_STAGE_IN_USE";
    pub const REQUEST_STAGE_FK: &str = "MAINTENANCEREQUEST_REQUEST_STAGE_FK";
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for MaintenanceRequest.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /maintenance_requests | List with pagination |
/// | POST | /maintenance_requests | Create new |
/// | GET | /maintenance_requests/:id | Get by ID |
/// | PUT | /maintenance_requests/:id | Full update |
/// | PATCH | /maintenance_requests/:id | Partial update |
/// | DELETE | /maintenance_requests/:id | Soft delete |
/// | POST | /maintenance_requests/bulk | Bulk create |
/// | POST | /maintenance_requests/upsert | Upsert |
/// | GET | /maintenance_requests/trash | List deleted |
/// | POST | /maintenance_requests/:id/restore | Restore |
/// | DELETE | /maintenance_requests/empty | Empty trash |
/// | GET | /maintenance_requests/:id/deleted | Get deleted by ID |
/// | DELETE | /maintenance_requests/trash/:id | Permanent delete from trash |
/// | GET | /maintenance_requests/count | Count active entities |
/// | GET | /maintenance_requests/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(MaintenanceRequestService::with_repository(repository));
/// let router = create_maintenance_request_routes(service);
/// ```
pub fn create_maintenance_request_routes(service: Arc<MaintenanceRequestService>) -> Router {
    BackboneCrudHandler::<MaintenanceRequestService, MaintenanceRequest, CreateMaintenanceRequestDto, UpdateMaintenanceRequestDto, MaintenanceRequestResponseDto>::routes(
        service,
        "/maintenance_requests",
    )
}

/// Create Axum router with only the read (GET) endpoints for MaintenanceRequest.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_maintenance_request_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_maintenance_request_read_routes(service: Arc<MaintenanceRequestService>) -> Router {
    BackboneCrudHandler::<MaintenanceRequestService, MaintenanceRequest, CreateMaintenanceRequestDto, UpdateMaintenanceRequestDto, MaintenanceRequestResponseDto>::read_routes(
        service,
        "/maintenance_requests",
    )
}

/// Create Axum router with only the write (mutation) endpoints for MaintenanceRequest.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_maintenance_request_write_routes(service: Arc<MaintenanceRequestService>) -> Router {
    BackboneCrudHandler::<MaintenanceRequestService, MaintenanceRequest, CreateMaintenanceRequestDto, UpdateMaintenanceRequestDto, MaintenanceRequestResponseDto>::write_routes(
        service,
        "/maintenance_requests",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_maintenance_request_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<MaintenanceRequestService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_maintenance_request_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}